pyproject.toml
[tool.poetry]
name = "wetoowrite"
version = "0.1.0"
description = ""
authors = ["rajesitb <56825534+rajesitb@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
dj-database-url = "0.5.0"
django-crispy-forms = "1.12.0"
django-widget-tweaks = "1.4.8"
Django = "3.2.12"
gunicorn = "^20.1.0"
whitenoise = {extras = ["brotli"], version = "^6.0.0"}
psycopg2 = "2.9.1"
boto3 = "^1.21.32"
django-countries = "7.3.2"
django-markdown-deux = "1.0.5"
django-ipware = "2.1.0"
django-redis = "4.12.1"
django-storages = "1.9.1"
django-taggit = "1.3.0"
tweepy = "^4.8.0"
textblob = "0.15.3"
django-tracking-analyzer = "1.0.2"
django-currentuser = "0.5.3"
better-profanity = "0.6.1"
django-ckeditor = "5.9.0"
Pillow = "7.1.2"
gensim = "3.8.3"
inflect = "0.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
